Add files via upload
It is a simple GUI based game which is designed using tkinter python libraries. Speaking about gameplay, it is a single player game where the player needs to answer some maths questions. The questions displayed are randomly chosen from a csv file. And there is a CSV file for the high score which is updated whenever the player beats the previous high score. We also used the pyttsx3 module to speak out the last 10 sec.
The player has to choose a level “easy” , “medium” or  “hard”.After clicking the level the game starts, the timer turns on and the question will be displayed. There will be 5 questions and every correct answer is worth 20 points. The player should answer the questions within the time limit.The limit varies over level for Easy:75 sec, Medium:90 sec Hard:150sec. The last 10 seconds will be counted to notify the player time. Either by answering all the questions or if  the time ends the game will reach to end. Click the Close button to close the game or Restart to play again.